==> R/10-maps-and-range-size-results.R <==
###############################################################################
## Project - CAR-SDM & climate change                                      ----
## Script  - 10-maps-and-range-size-results.R
## Updated - 01-30-2023
## Author  - Karin Kralicek (karin.kralicek@usda.gov)
## 
## This R script is in support of 'Spatial-Bayesian models project shifts in 
## suitable habitat for Pacific Northwest tree species under climate change' by
## Karin Kralicek, Jay M. Ver Hoef, Tara M. Barrett, and Temesgen Hailemariam.
## 
## About - this script:
## - Create maps and plots of mean-predictions and SE for the current and four
##   future scenarios. Figures correspond to:
##   - Manuscript: Figures 4 & 5 
##   - Appendix S1: Figures S2-S6 & S14-S17
##
## About - output:
## (Note to self: naming conventions differ from what I have saved (b/c the
##  figures moved around in the manuscript / appendices), what I have saved
##  things as is in parentheses after the names generated by the current
##  version of this script..)
## - ch2-10-map-base.rds (ch2-10-F2-map-base.rds)
##   - location: path_output
##   - data that is used in this script to plot
## - ch2-10-range_size_key.rds
##   - location: path_output
##   - associate MCMC samples with it's range-size (ha) so that it is easier
##     to pull the associated Voronoi polygon map to visualize what this value
##     for range-size actually looks like.
## - ch2-10-quantile-voronoi-maps-only.rdata
##   - location: path_output
##   - data for MCMC samples corresponding to the 0.05 & 0.95 quantiles of the
##     range-size estimates when calculated based on the Voronoi polygon method 
##     of range estimation. Contains two objects, one is the data (maps_base) 
##     and the other is a vector with the quantile values (quant_range)
###############################################################################
# data manipulation
library(magrittr)    # for set_rownames()
library(dplyr)       # for ... E V E R Y T H I N G ...
library(purrr)       # working with lists
library(reshape2)    # melt() and dcast()
library(stringr)     # for string help (e.g. str_detect in functions)

# plotting
library(ggplot2)
library(patchwork)
library(sf)

# == Paths =============================================================== ####
# path to Rdata
path_USB <- "G:/"
path_top <- paste0(path_USB, "Karin phd thesis files/ch2 - code output/")

path_data   <- paste0(path_top, "data/")
path_output <- paste0(path_top, "output/")

# path to functions script
path_fn <- paste0(path_top, "scripts - real data/")

# where to ggsave-ed figures go
path_images <- paste0(path_top, "images/")

# == Load data =========================================================== ####
# - fns + spatial data --------------------------------------------------- ####
# load fns
source(paste0(path_fn, "0-functions.R"))

# area data
plot_polygons_true.df <- readRDS(paste0(path_output, "ch2-8-plot_polygons_true_df.rds"))

# spatial plot mapping data
plot_polygons_public.sf <- readRDS(paste0(path_output, "ch2-8-plot_polygons_public_sf.rds"))

# study area boundary mapping data
study_area.sf <- readRDS(paste0(path_output, "ch2-8-study_area_sf.rds"))

# - plot key with lat & elevation ---------------------------------------- ####
elev_lat_key <- readRDS(paste0(path_data, "ch2-1-elev-lat-plot-key.rds"))

# == Arguments for figures =============================================== ####
scen_colPal <- data.frame(
  scenario = c("current",
               "CCSM4_rcp45_m2085",
               "HadGEM2_ES_rcp45_m2085",
               "CCSM4_rcp85_m2085",
               "HadGEM2_ES_rcp85_m2085"),
  # colorblind safe (cb) diverging palettes 
  # - reds (H) and blues (C)
  hex_cb_br = c("#999999", 
                "#56B4E9",  "#E69F00",  "#0072B2",  "#D55E00"),
  # - purple (H) and greens (C)
  # This is the one we are really going with...
  hex_cb_gp = c("#999999", 
                "#af8dc3",  "#7fbf7b", "#762a83",  "#1b7837"),
  hex_cb_mix = c("#999999",
                 "#56B4E9", "#E69F00", "#009E73", "#F0E442"),
  # - blue (C) and green (H)
  hex_cb_gb = c("#999999",
                "#a6cee3", "#b2df8a", "#1f78b4", "#33a02c")
)

# common names for species
spp_common <- c("noble fir",
                "coastal Douglas-fir",
                "blue oak",
                "white oak",
                "black oak") %>%
  setNames(c("abpr", "psmem", "qudo", "quga4", "quke"))

# set general text-size & figure width arguments (based on Ecosphere's author
# guidelines)
# - text: 6-10 pt
EJ_axis_text_size  <- 7
EJ_strip_text_size <- 7
EJ_axis_title_size <- 8
EJ_label_size      <- 10

# - max dimensions
EJ_dim_units <- "cm"
EJ_dim_1col_width_max <- 8.5 #cm
EJ_dim_2col_width_max <- 18  #cm
EJ_dim_2col_ht_max    <- 22  #cm

# - max dim appendices (b/c supplying a pdf of word doc with 1" margins)
EJ_app_dim_units <- "in"
EJ_app_dim_width_max <- 6.5 #in
EJ_app_dim_ht_max    <- 9   #in

# - dpi: min 300-600
EJ_dpi <- 800

# == Mean-prediction & uncertainty maps ================================== ####
# - (create|load) maps to plot ------------------------------------------- ####
# to save on space and time, pull the data we will plot and (re-)load that
# only into R...

# create or load the data...
if (file.exists(paste0(path_output, "ch2-10-map-base.rds"))) {
  map_base <- readRDS(paste0(path_output, "ch2-10-map-base.rds"))
} else {
  # (drop plot ids)
  drop_plot_ids <- readRDS(paste0(path_output, "ch2-8-drop_plot_ids.rds"))
  
  # (current climate predictions)
  preds_PRISM <- readRDS(paste0(path_output, "ch2-7-preds-PRISM.rds"))$preds_PRISM %>%
    imap(function(preds.x, sp_name) {
      preds.x$maps_2k %>% 
        map(~.x %>% select(plot_id, mu_hat, scenario)) %>%
        bind_rows() %>%
        filter(!plot_id %in% drop_plot_ids[[sp_name]]) %>%
        group_by(plot_id, scenario) %>%
        summarize(mu_meanpred = mean(mu_hat),
                  se_mu = sd(mu_hat))
    })
  
  # (future climate predictions differenced from current)
  future_scenarios <- readRDS(paste0(path_output, "ch2-7-future_scenarios_chr.rds"))
  preds_CNA <- lapply(future_scenarios, function(x) {
    # this is a list of species under one scenario
    readRDS(paste0(path_output, "ch2-7-preds-CNA_", x, ".rds"))$preds_CNA %>%
      imap(function(preds.x, sp_name) {
        preds.x$maps_2k %>% 
          map(~.x %>% select(plot_id, mu_hat, scenario)) %>%
          bind_rows() %>%
          filter(!plot_id %in% drop_plot_ids[[sp_name]]) %>%
          group_by(plot_id, scenario) %>%
          summarize(mu_meanpred = mean(mu_hat),
                    se_mu = sd(mu_hat))
      })
  }) %>%
    setNames(future_scenarios) %>%
    transpose
  preds_CNA <- list(preds_PRISM, preds_CNA) %>%
    pmap(function(preds_current.x, preds_future.x) {
      preds_future.x %>%
        map(function(scen.x) {
          # diff = future - current
          scen.x %>% 
            inner_join(preds_current.x %>% select(-scenario), by = "plot_id") %>%
            mutate(mu_meanpred = mu_meanpred.x - mu_meanpred.y,
                   se_mu = se_mu.x - se_mu.y) %>%
            select(plot_id, mu_meanpred, se_mu, scenario) %>%
            mutate(stat_type = "diff")
        }) %>%
        bind_rows
    })
  
  # (make final subsetted object)
  map_base <- list(preds_PRISM, preds_CNA) %>%
    pmap(function(preds_current.x, preds_future.x) {
      preds_current.x %>%
        select(plot_id, mu_meanpred, se_mu, scenario) %>%
        mutate(stat_type = "orig") %>%
        rbind.data.frame(preds_future.x)
    })
  
  # save objects
  saveRDS(map_base,
          file = paste0(path_output, "ch2-10-map-base.rds"))
}

# - maps: mean-prediction & SE (Figure 5 & Figures S14-S17) -------------- ####
# (future differenced from current)
# - these correspond to Figures that appear in the manuscript / appendix
p_body <- list(map_base %>% map(~.x %>% rename(mu_hat = mu_meanpred)), 
               names(map_base), 
               plot_polygons_public.sf) %>%
  # map(~.x[5]) %>% # to generate a map for only one species
  pmap(~fn.map_mp_uncert(..1, ..2, ..3,
                         study_area.sf, 
                         T_fnodiff_used = FALSE,
                         EJ_options_check = TRUE))

p_body %>%
  setNames(paste0("Fig", c("S14", "5", "S15", "S16", "S17"))) %>%
  imap(function(p.x, name.x) {
    if (name.x == "Fig5") {
      ggsave(filename = paste0(path_images, name.x, ".tiff"),
             plot = p.x,
             height = 4.5*(EJ_dim_2col_width_max/6),
             width = EJ_dim_2col_width_max,
             units = EJ_dim_units, 
             dpi = EJ_dpi,
             compression = "lzw",
             type = "cairo")
    } else {
      ggsave(filename = paste0(path_images, name.x, ".tiff"),
             plot = p.x,
             height = 4.5*(EJ_app_dim_width_max/6),
             width = EJ_app_dim_width_max,
             units = EJ_app_dim_units, 
             dpi = EJ_dpi,
             compression = "lzw",
             type = "cairo")
    }
  })

# (future not differenced from current - ie, on the probability-scale)
map_base.FnotDiff <- map_base %>% 
  map(function(x) {
    x %>% 
      filter(scenario != "current") %>%
      select(-stat_type) %>%
      inner_join(x %>% 
                   filter(scenario == "current") %>% 
                   select(-scenario) %>%
                   rename(mu_current = mu_meanpred,
                          se_mu_current = se_mu),
                 by = "plot_id") %>%
      mutate(mu_meanpred = mu_meanpred + mu_current,
             se_mu = se_mu + se_mu_current) %>%
      select(-mu_current, -se_mu_current) %>%
      rbind.data.frame(., x %>% filter(scenario == "current"))
  })

list(map_base.FnotDiff %>% map(~.x %>% rename(mu_hat = mu_meanpred)), 
     names(map_base.FnotDiff), 
     plot_polygons_public.sf) %>%
  map(~.x["abpr"]) %>% # to generate a map for only one species
  pmap(~fn.map_mp_uncert(..1, ..2, ..3, 
                         study_area.sf, 
                         T_fnodiff_used = TRUE))

# - plot: change in mu_hat (f-c) plotted by elev & lat (Figure 4) -------- ####
# relationship to elev & lat of diff in either (1) mean-pred or (2) SE
map_base %>%
  imap(function(sp.x, sp_name) {
    # create new labels for scenarios...
    new_lev <- c("RCP 4.5 CCSM4",
                 "RCP 4.5 HadGEM2-ES",
                 "RCP 8.5 CCSM4",
                 "RCP 8.5 HadGEM2-ES")
    old_lev <- c("CCSM4_rcp45_m2085",
                 "HadGEM2_ES_rcp45_m2085",
                 "CCSM4_rcp85_m2085",
                 "HadGEM2_ES_rcp85_m2085")
    # format data for plotting
    sp.x %>% 
      # ... grab the difference in mean-preds & SE...
      filter(stat_type == "diff") %>%
      rename(mu_hat_diff = mu_meanpred,
             se_mu_diff = se_mu) %>% 
      select(scenario, plot_id, mu_hat_diff, se_mu_diff) %>%
      melt(id.vars = c("scenario", "plot_id"),
           variable.name = "stat") %>%
      left_join(elev_lat_key %>% select(-LON_ACTUAL), by = "plot_id") %>%
      # ... associate new labels and make into a factor...
      mutate(scenario = case_when(scenario == old_lev[1] ~ new_lev[1],
                                  scenario == old_lev[2] ~ new_lev[2],
                                  scenario == old_lev[3] ~ new_lev[3],
                                  scenario == old_lev[4] ~ new_lev[4]),
             spp = recode(sp_name, !!!spp_common),
             spp = factor(spp, levels = unname(spp_common)),
             elev_km = Elev_m / 1000) %>%
      mutate(scenario = factor(scenario, levels = new_lev)) %>%
      select(-Elev_m) %>%
      split(.$stat)
  }) %>%
  transpose %>%
  map(~.x %>% bind_rows) %>%
  # make the figure
  imap(function(df.x, stat) {
    # set color based on stat --------------------------------------- ####
    TF_mp <- stat == "mu_hat_diff"
    
    # change color-scheme based on if diff is mean-preds or SE...
    f_cb_name   <- ifelse(TF_mp, "diff prob.", "diff SE")
    f_cb_low    <- ifelse(TF_mp, "#39568cff", "#451077FF")
    f_cb_high   <- ifelse(TF_mp, "#FDD835", "#FDD835")
    title_name  <- ifelse(TF_mp, "mean-prediction", "SE of prediction")
    
    # future plots -------------------------------------------------- ####
    ggplot(df.x) +
      stat_summary_hex(aes(y = LAT_ACTUAL, x = elev_km, z = value),
                       fun = ~ mean(.x),
                       bins = 40) +
      facet_grid(spp ~ scenario, scales = "free") +
      scale_fill_gradientn(
        name = f_cb_name,
        colors = c(f_cb_low, "#e5e5e5", f_cb_high),
        values = scales::rescale(c(-1, -0.25, 0, 0.25, 1)),
        limits = function(x) {c(-1,1)*max(abs(x))}) +
      theme_bw() +
      theme(aspect.ratio = 1,
            panel.grid = element_blank(),
            panel.background = element_rect(fill = "grey50")) + 
      labs(x = "Elevation (km)",
           y = "Latitude (°N)",
           title = paste0(title_name, " of future climate scenarios",
                          "\nshown as differenced from current value (f-c)"))
    
  }) %>%
  # print out only mp's atm...
  pluck(1)

# add current preds to plot...
# Description: 5 (spp, rows) x 5 (scenarios, cols) figure showing either
# the mean-preds (prob for current, diff prob for future scenarios) or SE
# (similarly, SE or diff SE for current or future scenarios, respectively)
map_base %>%
  # format data for plotting
  imap(function(sp.x, sp_name) {
    sp.x %>% 
      bind_rows() %>%
      # grab current (mp & se) & future (dif-mp & dif-se)
      mutate(dummy = paste0(scenario, "_", stat_type)) %>%
      filter(dummy == "current_orig" | stat_type == "diff") %>%
      # pretty labels for plotting
      mutate(scenario = case_when(
        scenario == "current" ~ "current\n",
        scenario == "CCSM4_rcp45_m2085" ~ "RCP 4.5\n CCSM4",
        scenario == "HadGEM2_ES_rcp45_m2085" ~ "RCP 4.5\n HadGEM2-ES",
        scenario == "CCSM4_rcp85_m2085" ~ "RCP 8.5\n CCSM4",
        scenario == "HadGEM2_ES_rcp85_m2085" ~ "RCP 8.5\n HadGEM2-ES")) %>%
      mutate(scenario = factor(scenario, 
                               levels = c("current\n",
                                          "RCP 4.5\n CCSM4",
                                          "RCP 4.5\n HadGEM2-ES",
                                          "RCP 8.5\n CCSM4",
                                          "RCP 8.5\n HadGEM2-ES"))) %>%
      tidyr::pivot_longer(cols = c("mu_meanpred", "se_mu"),
                          names_to = "stat") %>%
      left_join(elev_lat_key %>% select(-LON_ACTUAL), by = "plot_id") %>%
      mutate(spp = recode(sp_name, !!!spp_common),
             spp = factor(spp, levels = unname(spp_common)),
             elev_km = Elev_m / 1000) %>%
      select(-Elev_m) %>%
      split(.$stat)
  }) %>%
  transpose %>%
  map(~.x %>% bind_rows) %>%
  # make the figure
  imap(function(df.x, stat) {
    # set color based on stat --------------------------------------- ####
    TF_mp <- stat == "mu_meanpred"
    
    # change color-scheme based on if diff is mean-preds or SE...
    c_cb_name   <- ifelse(TF_mp, "prob.", "SE")
    c_cb_option <- ifelse(TF_mp, "viridis", "magma")
    f_cb_name   <- ifelse(TF_mp, "diff prob.", "diff SE")
    f_cb_low    <- ifelse(TF_mp, "#39568cff", "#451077FF")
    f_cb_high   <- ifelse(TF_mp, "#FDD835", "#FDD835")
    title_name  <- ifelse(TF_mp, "mean-prediction", "SE of prediction")
    
    # current plot -------------------------------------------------- ####
    p_c <- ggplot(df.x %>% filter(scenario == "current\n")) +
      stat_summary_hex(aes(y = LAT_ACTUAL, x = elev_km, z = value),
                       fun = ~ mean(.x),
                       bins = 40) +
      facet_grid(spp ~ scenario, scales = "free") +
      scale_fill_viridis_c(name = c_cb_name,
                           option = c_cb_option) +
      theme_bw()  +
      labs(y = "Latitude (°N)") +
      theme(aspect.ratio = 1,
            panel.grid = element_blank(),
            panel.background = element_rect(fill = "grey50"),
            strip.background.y = element_blank(),
            strip.text.y = element_blank(),
            axis.title.x = element_blank(),
            plot.margin = margin(1,0,1,.5),
            strip.text.x = element_text(size = EJ_strip_text_size),
            axis.title.y = element_text(size = EJ_axis_title_size),
            axis.text = element_text(size = EJ_axis_text_size,
                                     color = "black"),
            legend.title = element_text(size = EJ_axis_title_size),
            legend.text  = element_text(size = EJ_strip_text_size),
            panel.spacing = unit(.2, "lines"))
    
    # future plots -------------------------------------------------- ####
    # plot points on grey background for visibility
    p_f <- ggplot(df.x %>% filter(scenario != "current\n")) +
      stat_summary_hex(aes(y = LAT_ACTUAL, x = elev_km, z = value),
                       fun = ~ mean(.x),
                       bins = 40) +
      facet_grid(spp ~ scenario, scales = "free") +
      scale_fill_gradientn(
        name = f_cb_name,
        colors = c(f_cb_low, "#e5e5e5", f_cb_high),
        values = scales::rescale(c(-1, -0.25, 0, 0.25, 1)),
        limits = function(x) {c(-1,1)*max(abs(x))}) +
      theme_bw() + 
      labs(x = "Elevation (km)") +
      theme(aspect.ratio = 1,
            panel.grid = element_blank(),
            panel.background = element_rect(fill = "grey50"),
            axis.title.y = element_blank(),
            axis.text.y = element_blank(),
            axis.ticks.y = element_blank(),
            plot.margin = margin(1,.5,1,0),
            axis.title.x = element_text(hjust = 0.35,
                                        size = EJ_axis_title_size),
            strip.text.x = element_text(size = EJ_strip_text_size),
            strip.text.y = element_text(size = EJ_strip_text_size),
            axis.text.x  = element_text(size = EJ_axis_text_size,
                                        color = "black"),
            legend.title = element_text(size = EJ_axis_title_size),
            legend.text  = element_text(size = EJ_strip_text_size),
            panel.spacing = unit(.2, "lines")) 
    
    # # combine into one figure & return ------------------------------ ####
    # p_title <- paste0("Pairwise difference between climate data sets' ",
    #                   title_name,
    #                   "/nmean displayed for plots within binned area")
    # 
    # (p_c | p_f) + 
    #   plot_layout(widths = c(1.05, 4),
    #               heights = c(1,1),
    #               guides = "collect") & 
    #   theme(legend.position = "bottom")
    list(p_c, p_f)
    
  }) %>%
  # tweak names
  setNames(c("Fig4", "plot_elevlat_se_mu_allsppscenarios")) %>%
  # ggsave the plots
  imap(function(p.x, name.x) {
    ggsave(filename = paste0(path_images, name.x, ".tiff"),
           plot = (p.x[[1]] | p.x[[2]]) + 
             plot_layout(widths = c(1.05, 4),
                         heights = c(1,1),
                         guides = "collect") &
             theme(legend.position = "bottom"),
           height = 6.5*(EJ_dim_2col_width_max/6), 
           width = EJ_dim_2col_width_max,
           units = EJ_dim_units, 
           dpi = EJ_dpi,
           compression = "lzw")
  })


# == Range-size est (VP method) ========================================== ####
# - (create|load) range_size key ----------------------------------------- ####
# Create a list of dfs by spp containing range-size estimates for each MCMC
# sample for each scenario (also calculate value if based on mean-predictions)
# - this takes a while, so create and then (re-)load.
# - here, range-size is estimated based on the Voronoi polygon method of range
#   estimation. 
if (file.exists(paste0(path_output, "ch2-10-range_size_key.rds"))) {
  range_size_key <- readRDS(paste0(path_output, "ch2-10-range_size_key.rds"))
} else {
  # define function to associate range-size (ha) with it's MCMC_iter
  fn.range_size_df <- function(preds.x, poly.x, drop_ids) {
    # Variables:
    # - area (units ha) is based on Voronoi polygons (see script 8 for calc);
    # - range-size is the sum of area_ha for all plots where pps_pa is 'P'
    preds.x[c("maps_mean", "maps_2k")] %>%
      modify_at("maps_mean", function(x) {
        pps_p_plots <- x %>% 
          filter(!plot_id %in% drop_ids) %>%
          arrange(plot_id) %>% 
          pull(pps_pa) %>% 
          as.logical
        
        range_size_ha <- sum((poly.x %>% 
                                filter(!plot_id %in% drop_ids) %>% 
                                arrange(plot_id))[pps_p_plots, ]$area_ha)
        
        data.frame(map_group = "maps_mean",
                   range_size_ha = range_size_ha,
                   MCMC_iter = 1)           
        
      }) %>%
      modify_at("maps_2k", function(x_top) {
        x_top %>% 
          map(function(x) {
            pps_p_plots <- x %>% 
              filter(!plot_id %in% drop_ids) %>%
              arrange(plot_id) %>% 
              pull(pps_pa) %>% 
              as.logical
            
            range_size_ha <- sum((poly.x %>% 
                                    filter(!plot_id %in% drop_ids) %>% 
                                    arrange(plot_id))[pps_p_plots, ]$area_ha)
            
            data.frame(map_group = "maps_2k",
                       range_size_ha = range_size_ha)           
            
          }) %>%
          bind_rows %>%
          mutate(MCMC_iter = 1:n())
      }) %>%
      bind_rows
  }
  
  # drop plot ids
  drop_plot_ids <- readRDS(paste0(path_output, "ch2-8-drop_plot_ids.rds"))
  
  # preds_PRISM (current climate predictions)
  preds_PRISM <- readRDS(paste0(path_output, "ch2-7-preds-PRISM.rds"))$preds_PRISM
  # - get range-sizes only
  range_size.PRISM <- list(preds_PRISM, plot_polygons_true.df, drop_plot_ids) %>% 
    pmap(~fn.range_size_df(..1, ..2, ..3))
  
  # preds_CNA (future climate predictions)
  # - names of future scenarios
  future_scenarios <- readRDS(paste0(path_output, "ch2-7-future_scenarios_chr.rds"))
  # - read-in predictions!
  preds_CNA <- lapply(future_scenarios, function(x) {
    readRDS(paste0(path_output, "ch2-7-preds-CNA_", x, ".rds"))#$preds_CNA
  }) %>%
    setNames(future_scenarios)
  # - get range-sizes only
  range_size.CNA <- preds_CNA %>% 
    map(~list(.x, plot_polygons_true.df, drop_plot_ids) %>%
          pmap(~fn.range_size_df(..1, ..2, ..3)))
  
  # merge the results together..
  range_size_key <- append(range_size.CNA, list(range_size.PRISM)) %>%
    setNames(c(future_scenarios, "current")) %>% 
    transpose %>% 
    map(~.x %>% bind_rows(.id = "scenario"))
  
  # clean-up space
  rm(range_size.CNA, range_size.PRISM, fn.range_size_df)
  
  # save key
  saveRDS(range_size_key,
          file = paste0(path_output, "ch2-10-range_size_key.rds"))
}

# - (create|load) quantiles & maps to plot ------------------------------- ####
# create two objects:
# - maps_base: list by spp, where each element is a list of length 2. Those
#   two elements correspond to the dfs for the MCMC samples that have the lower
#   (0.05) and upper (0.95) quantile of range-size estimates (based on the VP 
#   range estimation method). (unfortunate name similarity)
# - quant_range: hard set atm to c(.05, .95) for reference
if (file.exists(paste0(path_output, 
                       "ch2-10-quantile-voronoi-maps-only.rdata"))) {
  load(paste0(path_output, "ch2-10-quantile-voronoi-maps-only.rdata"))
} else {
  # upper and lower quantiles will be used as vlines in hist/dens of
  # range-size, and to pluck out corresponding maps for visualization
  quant_range <- c(.05, .95)
  
  # identify maps based on upper and lower quantiles (2k maps) range-size
  map_ids <- range_size_key %>% 
    map(function(key.x, q_range = quant_range) {
      # identify the quantile location to pull
      q_value <- quantile(1:max(key.x$MCMC_iter), q_range) %>%
        as.numeric %>%
        round(0)
      
      # grab correct map ids
      key.x %>%
        filter(map_group == "maps_2k") %>%
        group_by(scenario) %>%
        arrange(range_size_ha) %>%
        mutate(range_rank = 1:n()) %>%
        ungroup %>%
        filter(range_rank %in% q_value) %>%
        mutate(q_type = ifelse(range_rank == q_value[1],
                               paste0("q_", q_range[1]*100),
                               paste0("q_", q_range[2]*100))) %>%
        select(scenario, MCMC_iter, range_size_ha, q_type)
    })
  
  # extract only maps that will be plotted
  preds_PRISM <- readRDS(paste0(path_output, "ch2-7-preds-PRISM.rds"))$preds_PRISM
  future_scenarios <- readRDS(paste0(path_output, "ch2-7-future_scenarios_chr.rds"))
  preds_CNA <- lapply(future_scenarios, function(x) {
    readRDS(paste0(path_output, "ch2-7-preds-CNA_", x, ".rds"))#$preds_CNA
  }) %>%
    setNames(future_scenarios)
  
  # (drop plot ids)
  drop_plot_ids <- readRDS(paste0(path_output, "ch2-8-drop_plot_ids.rds"))
  
  maps_base <- 
    list(append(preds_CNA, list(preds_PRISM)) %>%
           setNames(c(future_scenarios, "current")) %>% 
           transpose,
         map_ids, drop_plot_ids) %>% 
    pmap(function(preds_ls, map_ids.x, drop_ids) {
      preds_ls %>% 
        imap(function(preds.x, scen_name.x) {
          ids.x <- map_ids.x %>% filter(scenario == scen_name.x)
          
          lapply(ids.x$MCMC_iter, function(i) {
            preds.x$maps_2k[[i]] %>% 
              filter(!plot_id %in% drop_ids) %>%
              mutate(range_size_ha = ids.x[ids.x$MCMC_iter == i, ]$range_size_ha,
                     q_type = ids.x[ids.x$MCMC_iter == i, ]$q_type)
          })
        })
    })
  
  save(maps_base, quant_range,
       file = paste0(path_output, "ch2-10-quantile-voronoi-maps-only.rdata"))
}

# - posteriors of range-size (density/boxplots) -------------------------- ####
# Notes
# - these figures just for my own inspection; Appendix S3 Figure S1 is
#   produced by script 11-...R
# - (!) center line is the mean in patchwork code below
list(
  # density plots
  fn.plot_rangesize(range_size_key,
                    quant_range = quant_range,
                    plot_type = "density", 
                    scen_colPal = scen_colPal),
  # boxplots
  fn.plot_rangesize(range_size_key,
                    quant_range = quant_range,
                    plot_type = "boxplot", 
                    scen_colPal = scen_colPal) %>%
    map(~.x + 
          theme(axis.text.x = element_blank()) + 
          guides(color = "none", fill = "none"))) %>%
  pmap(~ .x + .y) %>%
  reduce(., `/`) +
  plot_layout(guides = "collect") +
  plot_annotation(title = "Voronoi range-size estimates",
                  subtitle = "Based on 2k MCMC samples",
                  caption = "Note: abpr visualized on logp1 scale (ln(x + 1))") & 
  theme(legend.position = "bottom")

# density plots
fn.plot_rangesize(range_size_key,
                  quant_range = quant_range,
                  plot_type = "density", 
                  scen_colPal = scen_colPal) %>%
  reduce(., `/`) +
  plot_layout(guides = "collect") +
  plot_annotation(title = "Voronoi range-size estimates",
                  subtitle = "Based on 2k MCMC samples",
                  caption = "Note: abpr visualized on logp1 scale (ln(x + 1))")

# boxplots
fn.plot_rangesize(range_size_key,
                  quant_range = quant_range,
                  plot_type = "boxplot", 
                  scen_colPal = scen_colPal) %>%
  reduce(., `/`) +
  plot_layout(guides = "collect") +
  plot_annotation(title = "Voronoi range-size estimates",
                  subtitle = "Based on 2k MCMC samples, {min, q_05, mean, q_95, max}",
                  caption = "Note: abpr visualized on logp1 scale (ln(x + 1))") & 
  theme(legend.position = "bottom")

# - posteriors of change in range-size (summary/density) ----------------- ####
# Summary
# - by species and future climate scenario, equal-tailed 90% credible intervals 
#   for the difference in estimated range-size (thousand km sq) between that 
#   future climate scenario and the current scenario (as future - current).
range_size_key %>% map(function(sp.x) {
   sp.x %>%
      filter(map_group == "maps_2k") %>%
      # summarize the difference in range-size estimate by MCMC_iter...
      split(.$MCMC_iter) %>%
      map(function(x) {
        current_est <- x[x$scenario == "current", ]$range_size_ha
        x %>% 
          filter(scenario != "current") %>%
          mutate(range_diff_ha = range_size_ha - current_est) %>%
          mutate(range_diff_ksqkm = (range_diff_ha/100)/1000) %>%
          select(scenario, range_diff_ksqkm, MCMC_iter)
      }) %>%
      bind_rows() %>%
      group_by(scenario) %>%
      summarize(q05_diff_ksqkm = quantile(range_diff_ksqkm, 0.05) %>% unname,
                q95_diff_ksqkm = quantile(range_diff_ksqkm, 0.95) %>% unname)
  })

# density plots
# - posterior distributions for the difference (f - c) in range-size estimates
#   between the current and a given future scenario:
range_size_key %>%
  imap(function(sp.x, sp_name) {
    # new scenario legend titles for plotting
    legend_values <- with(scen_colPal, hex_cb_gp %>% setNames(scenario))
    legend_labels <- legend_values %>%
      names %>%
      gsub("_m2085", "", .) %>% 
      gsub("_rcp", " rcp", .) %>% 
      str_split(., " ") %>% 
      lapply(function(x) paste0(x[2], " ", x[1])) %>% 
      unlist %>% gsub("NA ", "", .)
    names(legend_values) <- legend_labels
    
    sp.x <- sp.x %>%
      filter(map_group == "maps_2k") %>%
      # add new scenario legend titles for plotting
      rowwise %>%
      mutate(scenario = legend_labels[grep(scenario, scen_colPal$scenario)]) %>% 
      mutate(scenario = factor(scenario, levels = legend_labels)) %>%
      ungroup %>%
      # summarize the difference in range-size estimate by MCMC_iter...
      split(.$MCMC_iter) %>%
      map(function(x) {
        current_est <- x[x$scenario == "current", ]$range_size_ha
        x %>% 
          filter(scenario != "current") %>%
          mutate(range_diff_ha = range_size_ha - current_est) %>%
          mutate(range_diff_ksqkm = (range_diff_ha/100)/1000) %>%
          select(scenario, range_diff_ksqkm, MCMC_iter) %>%
          mutate(spp = toupper(sp_name))
      }) %>%
      bind_rows()
    
    legend_labels <- legend_labels[2:5]
    legend_values <- legend_values[2:5]
    
    # make the plots
    if (sp_name == "quke") {
      ggplot(sp.x %>% filter(spp != "abpr")) +
        geom_density(aes(x = range_diff_ksqkm, 
                         y = ..density.., 
                         fill = scenario),
                     alpha = .5,
                     trim = TRUE) +
        facet_grid(spp ~ ., scales = "free_y") +
        scale_color_manual(values = legend_values, guide = FALSE) +
        scale_fill_manual(values = legend_values,
                          labels = gsub(" ", "\n", legend_labels))  +
        xlab("diff. total range-size") +
        ylab("density") +
        theme_bw() +
        guides(colour = guide_legend(override.aes = list(alpha = 1))) +
        theme(plot.margin = margin(0, 0, 5.5, 0),
              legend.title = element_blank())
    } else {
      ggplot(sp.x) +
        geom_density(aes(x = range_diff_ksqkm, 
                         y = ..density.., 
                         fill = scenario),
                     alpha = .5,
                     trim = TRUE) +
        facet_grid(spp ~ ., scales = "free_y") +
        scale_color_manual(values = legend_values, guide = FALSE) +
        scale_fill_manual(values = legend_values,
                          labels = gsub(" ", "\n", legend_labels))  +
        xlab("diff. total range-size") +
        ylab("density") +
        theme_bw() +
        theme(plot.margin = margin(0, 0, 5.5, 0),
              axis.title.x = element_blank()) + 
        guides(color = "none", fill = "none")
    }
  }) %>%
  map(~.x + geom_vline(aes(xintercept = 0), color = "black")) %>%
  reduce(., `/`) +
  plot_layout(guides = "collect") +
  plot_annotation(title = "Posterior for Voronoi range-size change (thousand km^2)",
                  subtitle = "Based on 2k MCMC samples")

# - maps: lower/upper quant. of range-sizes est.(Figures S2-S6) ---------- ####
# Note: scen_colPal.hex argument is used only to order future scenario names
# (future probs & 2k km^2 differenced from current)
p_body <- list(maps_base, names(maps_base), plot_polygons_public.sf) %>%
  # map(~.x[1]) %>% # to generate a map for only one spp
  pmap(~fn.map_rangesize( ..1, ..2, ..3, study_area.sf, 
                          diff_map_T = TRUE))

# save figures S2-S6
p_body %>%
  setNames(paste0("FigS", 2:6)) %>%
  imap(function(p.x, name.x) {
    ggsave(filename = paste0(path_images, name.x, ".tiff"),
           plot = p.x,
           height = 4.91 * (EJ_app_dim_width_max/6), 
           width = EJ_app_dim_width_max,
           units = EJ_app_dim_units, 
           dpi = EJ_dpi,
           compression = "lzw",
           type = "cairo")
  })

# map the difference between future upper & lower quantiles
list(maps_base, names(maps_base), plot_polygons_public.sf) %>%
  map(~.x[2]) %>% # to generate a map for only one spp
  pmap(function(preds.x, sp_name.x, plot_sf.x) {
    # - make new legend labels --------------------------------------------- ####
    legend_labels <- with(scen_colPal, hex_cb_gp %>% setNames(scenario)) %>%
      names %>%
      gsub("_m2085", "", .) %>% 
      gsub("_rcp", " rcp", .) %>% 
      str_split(., " ") %>% 
      lapply(function(x) paste0(x[2], " ", x[1])) %>% 
      unlist %>% gsub("NA ", "", .)
    
    # - make the figure ---------------------------------------------------- ####
    (list(preds.x$current,
          preds.x[names(preds.x) != "current"] %>% transpose) %>%
       # for each quant do the following...
       pmap(function(current.x, future_ls.x) { 
         future_ls.x %>% 
           map(function(future.x) {
             future.x %>%
               inner_join(current.x %>% select(plot_id, mu_hat, range_size_ha),
                          by = "plot_id") %>%
               mutate(pred_diff = mu_hat.x - mu_hat.y) %>%
               select(plot_id, pred_diff, q_type)
           }) %>% 
           bind_rows(.id = "scenario") 
       }) %>%
       bind_rows %>%
       dcast(scenario + plot_id ~ q_type, value.var = "pred_diff") %>%
       group_by(scenario, plot_id) %>%
       mutate(pred_diff_diff = q_95 - q_5) %>%
       ungroup %>%
       # make scenario & quant names pretty...
       inner_join(data.frame(scenario = scen_colPal$scenario,
                             tmp = legend_labels),
                  by = "scenario") %>%
       select(-scenario) %>%
       rename(scenario = tmp) %>%
       mutate(scenario = factor(scenario, levels = legend_labels)) %>%
       inner_join(plot_sf.x %>% select(plot_id, geometry),
                  by = "plot_id")) %>%
      ggplot(aes(geometry = geometry))+
      geom_sf(data = study_area.sf) + 
      geom_sf(aes(fill = pred_diff_diff), color = NA) +
      facet_grid(. ~ scenario) +
      scale_fill_gradient2("diff of/ndiff prob.",
                           low = "red",
                           mid = "white",
                           high = "black",
                           midpoint = 0) +
      scale_x_continuous(breaks = c(-122, -118, -114)) +
      theme_bw()
  })
